git-svn-id: https://scm.mvapich.cse.ohio-state.edu/svn/mpi/mvapich2/trunk@6535 09bc9535-d30e-0410-b1f7-d46b20a4725c

+++ src/mpid/ch3/channels/mrail/src/gen2/ibv_ud.c
@@ -37,7 +37,7 @@ static inline void mv2_ud_flush_ext_window(MPIDI_VC_t *vc)
         q->tail = NULL;
     }
 }
-static inline void mv2_ud_process_ack(MPIDI_VC_t *vc, int acknum)
+static inline void mv2_ud_process_ack(MPIDI_VC_t *vc, uint16_t acknum)
 {
     vbuf *sendwin_head = vc->mrail.ud.send_window.head;
